[ACH] : correction tous les controlleurs

+++ src/main/java/com/pfa/ecommerce/repository/ArticleRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import com.pfa.ecommerce.entities.ArticleEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArticleRepository extends JpaRepository<ArticleEntity, Integer> {
+public interface ArticleRepository extends JpaRepository<ArticleEntity, Long> {
 
 
     //Optional<ArticleEntity> findByCodeArticle(Long codeArticle);
