+++ src/main/java/com/pfa/ecommerce/repository/DetailsRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 import java.util.Optional;
 
-public interface DetailsRepository extends JpaRepository<DetailsEntity,Integer> , JpaSpecificationExecutor<DetailsEntity> {
+public interface DetailsRepository extends JpaRepository<DetailsEntity,Long> , JpaSpecificationExecutor<DetailsEntity> {
 
 }
