+++ src/main/java/com/pfa/ecommerce/repository/DevisRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import com.pfa.ecommerce.entities.DevisEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DevisRepository extends JpaRepository<DevisEntity,Integer>
+public interface DevisRepository extends JpaRepository<DevisEntity,Long>
 {
 	//Optional<DevisEntity> findById(long CodeDevis);
